refactor: type 관련 파일로 옮기기
